pyproject.toml
[project]
name = "anonchik-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiobotocore>=2.21.1",
    "aiofiles>=24.1.0",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "greenlet>=3.1.1",
    "httpx>=0.28.1",
    "pillow>=11.2.1",
    "pydantic-settings>=2.8.1",
    "pyjwt[crypto]>=2.10.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]
